refactor: move convertType function
