+++ packages/serverless-typespec-generator/src/ir/build.test.ts
@@ -3,8 +3,8 @@ import { describe, expect, it, vi } from "vitest"
 import { formatTypeSpec } from "@typespec/compiler"
 import type Serverless from "serverless"
 import type { SLS } from "./../types/serverless"
-import type { JSONSchema, TypeSpecIR } from "./../typespec/ir/type"
-import { buildIR, jsonSchemaToTypeSpecIR } from "./build"
+import type { JSONSchema, PropTypeIR, TypeSpecIR } from "./../typespec/ir/type"
+import { buildIR, convertType, jsonSchemaToTypeSpecIR } from "./build"
 
 const context = describe
 
@@ -622,5 +622,40 @@ describe("jsonSchemaToTypeSpecIR", () => {
         },
       },
     })
+  })
+})
+
+describe("convertType", () => {
+  it("should convert string type to string", () => {
+    const schema: JSONSchema = { type: "string" }
+    const result = convertType(schema)
+    expect(result).toBe("string")
+  })
+  it("should convert integer type to numeric", () => {
+    const schema: JSONSchema = { type: "integer" }
+    const result = convertType(schema)
+    expect(result).toBe("numeric")
+  })
+  it("should convert number type to numeric", () => {
+    const schema: JSONSchema = { type: "number" }
+    const result = convertType(schema)
+    expect(result).toBe("numeric")
+  })
+  it("should convert boolean type to boolean", () => {
+    const schema: JSONSchema = { type: "boolean" }
+    const result = convertType(schema)
+    expect(result).toBe("boolean")
+  })
+  it("should convert array type to array of string", () => {
+    const schema: JSONSchema = { type: "array", items: { type: "string" } }
+    const result = convertType(schema)
+    expect(result).toEqual(["string"])
+  })
+  it("should convert oneOf type to union of types", () => {
+    const schema: JSONSchema = {
+      oneOf: [{ type: "string" }, { type: "null" }],
+    }
+    const result = convertType(schema)
+    expect(result).toEqual<PropTypeIR>({ union: ["string", "null"] })
   })
 })
