+++ packages/serverless-typespec-generator/src/ir/build.ts
@@ -1,13 +1,13 @@
 import type Aws from "serverless/aws"
 import { Registry } from "./../registry"
 import type { SLS } from "./../types/serverless"
-import { convertType } from "./../typespec/ir/convert"
 import { NotImplementedError } from "./../typespec/ir/error"
 import type {
   HttpResponseIR,
   JSONSchema,
   OperationIR,
   PropIR,
+  PropTypeIR,
   TypeSpecIR,
 } from "./../typespec/ir/type"
 
@@ -249,4 +249,35 @@ function mergeAllOfObjectSchemas(allOf: JSONSchema[]): Record<string, PropIR> {
     }
   }
   return props
+}
+
+export function convertType(schema: JSONSchema): PropTypeIR {
+  if (schema.oneOf) {
+    const types = schema.oneOf.map(convertType)
+    return { union: types }
+  }
+
+  if (schema.type === "object" || schema.allOf) {
+    return extractProps(schema)
+  }
+
+  switch (schema.type) {
+    case "string":
+      return "string"
+    case "integer":
+      return "numeric"
+    case "number":
+      return "numeric"
+    case "boolean":
+      return "boolean"
+    case "null":
+      return "null"
+    case "array":
+      if (!schema.items || Array.isArray(schema.items)) {
+        throw new Error("Array 'items' must be a single schema object")
+      }
+      return [convertType(schema.items)]
+    default:
+      throw new Error(`Unknown type: ${schema.type}`)
+  }
 }
