--- packages/serverless-typespec-generator/src/typespec/ir/convert.test.ts
@@ -1,39 +1,0 @@
-import { describe, expect, it } from "vitest"
-
-import { convertType } from "./convert"
-import type { JSONSchema, PropTypeIR } from "./type"
-
-describe("convertType", () => {
-  it("should convert string type to string", () => {
-    const schema: JSONSchema = { type: "string" }
-    const result = convertType(schema)
-    expect(result).toBe("string")
-  })
-  it("should convert integer type to numeric", () => {
-    const schema: JSONSchema = { type: "integer" }
-    const result = convertType(schema)
-    expect(result).toBe("numeric")
-  })
-  it("should convert number type to numeric", () => {
-    const schema: JSONSchema = { type: "number" }
-    const result = convertType(schema)
-    expect(result).toBe("numeric")
-  })
-  it("should convert boolean type to boolean", () => {
-    const schema: JSONSchema = { type: "boolean" }
-    const result = convertType(schema)
-    expect(result).toBe("boolean")
-  })
-  it("should convert array type to array of string", () => {
-    const schema: JSONSchema = { type: "array", items: { type: "string" } }
-    const result = convertType(schema)
-    expect(result).toEqual(["string"])
-  })
-  it("should convert oneOf type to union of types", () => {
-    const schema: JSONSchema = {
-      oneOf: [{ type: "string" }, { type: "null" }],
-    }
-    const result = convertType(schema)
-    expect(result).toEqual<PropTypeIR>({ union: ["string", "null"] })
-  })
-})
